# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис хранения логинов, паролей, бинарных данных и прочей приватной информации.",
        "title": "GophKeeper API",
        "contact": {},
        "version": "0.0.1"
    },
    "host": "0.0.0.0:8080",
    "basePath": "/api/v1",
    "paths": {
        "/all": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "All"
                ],
                "summary": "Получить все расшифрованные данные пользователя",
                "operationId": "all",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presentation.GetAllResponse"
                        }
                    },
                    "401": {
                        "description": "Нет токена авторизации, либо токен невалиден"
                    }
                }
            }
        },
...
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Авторизация пользователя по логину и паролю",
                "operationId": "auth-login",
                "parameters": [
                    {
                        "description": "Логин и Пароль",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presentation.registrationPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                            "Authorization eyJhbGciOiJI...qIScZUU8P0Zhck": {
                                "type": "string",
                                "description": "JWT"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный формат данных"
                    },
                    "401": {
                        "description": "Неправильный логин или пароль"
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация нового пользователя по логину и паролю",
                "operationId": "auth-register",
                "parameters": [
                    {
                        "description": "Логин и Пароль",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presentation.registrationPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                            "Authorization eyJhbGciOiJI...qIScZUU8P0Zhck": {
                                "type": "string",
                                "description": "JWT"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный формат данных"
                    },
                    "409": {
                        "description": "Логин уже занят"
                    }
                }
            }
        },
        "/bank_card/all": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "BankCard"
                ],
                "summary": "Получить все расшифрованные банковские карты",
                "operationId": "bank-card-all",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presentation.GetAllBankCardsResponse"
                        }
                    },
                    "401": {
                        "description": "Нет токена авторизации, либо токен невалиден"
                    }
                }
            }
        },
        "/bank_card/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "BankCard"
                ],
                "summary": "Создать и зашифровать банковскую карту",
                "operationId": "bank-card-create",
                "parameters": [
                    {
                        "description": "Номер, срок действия, cvv код, ФИО держателя карты",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presentation.bankCardPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "headers": {
                            "Location 020cb30c-c495-4a18-ac09-fd68c6f7c941": {
                                "type": "string",
                                "description": "UUID ресурса"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный формат данных"
                    },
                    "401": {
                        "description": "Нет токена авторизации, либо токен невалиден"
                    }
                }
            }
        },
        "/bank_card/{bank_card_id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "BankCard"
                ],
                "summary": "Обновить и зашифровать существующую банковскую карту",
                "operationId": "bank-card-update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bank Card ID",
                        "name": "bank_card_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Номер, срок действия, cvv код, ФИО держателя карты",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presentation.bankCardPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Некорректный формат данных или идентификатора"
                    },
                    "401": {
                        "description": "Нет токена авторизации, либо токен невалиден"
                    },
                    "404": {
                        "description": "Не найдено"
                    }
                }
            }
        },
        "/binary/all": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Binary"
                ],
                "summary": "Получить все расшифрованные бинарные данные",
                "operationId": "binary-all",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presentation.GetAllBinariesResponse"
                        }
                    },
                    "401": {
                        "description": "Нет токена авторизации, либо токен невалиден"
                    }
                }
            }
        },
        "/binary/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Binary"
                ],
                "summary": "Создать и зашифровать бинарные данные",
                "operationId": "binary-create",
                "parameters": [
                    {
                        "description": "Содержимое файла",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "headers": {
                            "Location 020cb30c-c495-4a18-ac09-fd68c6f7c941": {
                                "type": "string",
                                "description": "UUID ресурса"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный формат данных"
                    },
                    "401": {
                        "description": "Нет токена авторизации, либо токен невалиден"
                    }
                }
            }
        },
        "/binary/{binary_id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Binary"
                ],
                "summary": "Обновить и зашифровать существующие бинарные данные",
                "operationId": "binary-update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Binary ID",
                        "name": "binary_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Содержимое файла",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Некорректный формат данных или идентификатора"
                    },
                    "401": {
                        "description": "Нет токена авторизации, либо токен невалиден"
                    },
                    "404": {
                        "description": "Не найдено"
                    }
                }
            }
        },
        "/credentials/all": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Credentials"
                ],
                "summary": "Получить все расшифрованные логины и пароли",
                "operationId": "credentials-all",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presentation.GetAllCredentialsResponse"
                        }
                    },
                    "401": {
                        "description": "Нет токена авторизации, либо токен невалиден"
                    }
                }
            }
        },
        "/credentials/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Credentials"
                ],
                "summary": "Создать и зашифровать логин и пароль",
                "operationId": "credentials-create",
                "parameters": [
                    {
                        "description": "Наименование, логин и пароль",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presentation.credentialsPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "headers": {
                            "Location 020cb30c-c495-4a18-ac09-fd68c6f7c941": {
                                "type": "string",
                                "description": "UUID ресурса"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный формат данных"
                    },
                    "401": {
                        "description": "Нет токена авторизации, либо токен невалиден"
                    }
                }
            }
        },
...
        "/text/all": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Text"
                ],
                "summary": "Получить все расшифрованные текстовые данные",
                "operationId": "text-all",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presentation.GetAllTextsResponse"
                        }
                    },
                    "401": {
                        "description": "Нет токена авторизации, либо токен невалиден"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "presentation.GetAllBankCardsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "bank_cards": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/presentation.bankCardResponse"
                            }
                        }
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "presentation.registrationPayload": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}